Reuse ResponsePanel in Sidebar instead of a custom one. FIx an NPE
