+++ src/main/resources/com/javierllorente/netbeans/rest/client/HTTPExample.http
@@ -1,3 +1,5 @@
+https://jsonplaceholder.typicode.com/todos/1
+
 ### Request 1
 
 // POST request with json payload
